+++ package/tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(PSSDplusplus)
+
+test_check("PSSDplusplus")